Add files via upload
This is a report analysing data on which areas within the Sydney metropolitan area are the richest and poorest. This report used jupyter notebook, which sifted, cleaned, and consolidated the data at an early stage, employing SQL to sift and summarise the data.FinallyIt finally presentsThe final presentation is aFinally a report is presentedFinally a report is presented.Finally one is presented.It finally presentsThe final presentation is aThe final presentation is a completeFinally a complete data analysis is presentedFinally a complete data analysis report is presentedFinally a complete data analysis report is presented.